Se corrige un detalle en las facturas
Se corrige en las facturas para que los valores, en caso de ser un valor double con coma, se redondee a un numero entero.
Se corrige en el envio de mails para que se puedan poner lenguaje html en los mails de envio de bienvenida

+++ zapateria/src/main/java/persistencia/dao/mysql/DetalleFacturaDAOSQL.java
@@ -27,7 +27,7 @@ public class DetalleFacturaDAOSQL implements DetalleFacturaDAO{
 
 			statement.setInt(1, detalleFactura.getId());
 			statement.setInt(2, detalleFactura.getIdProducto());
-			statement.setInt(3, detalleFactura.getCantidad());
+			statement.setDouble(3, detalleFactura.getCantidad());
 			statement.setString(4, detalleFactura.getDescripcion());
 			statement.setDouble(5, detalleFactura.getPrecioCosto());
 			statement.setDouble(6, detalleFactura.getPrecioVenta());
@@ -72,7 +72,7 @@ public class DetalleFacturaDAOSQL implements DetalleFacturaDAO{
 	{
 		int id = resultSet.getInt("Id");
 		int idProducto = resultSet.getInt("IdProducto");
-		int cantidad = resultSet.getInt("Cantidad");
+		double cantidad = resultSet.getDouble("Cantidad");
 		String descripcion = resultSet.getString("Descripcion");
 		double precioCosto = resultSet.getDouble("PrecioCosto");
 		double precioVenta = resultSet.getDouble("PrecioVenta");
